Refactor to separate rtcAnalytics slice
